skeleting of javascripts in core
