+++ testjs/src/core/xajax.responseProcessor.js
@@ -1,0 +1,6 @@
+/*
+	Class: xajax.responseProcessor
+*/
+(function (xjx) {
+    xjx.responseProcessor = {};
+}(xajax));
